requirements.txt
click==8.1.3
Flask==2.2.3
future==0.18.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-whois==0.8.0
Werkzeug==2.2.3
